Various fixes

+++ Treblle/Constants.cs
@@ -7,7 +7,7 @@ namespace Treblle.Net
         /// <summary>
         /// SDK Payload Version - increment this number (21, 22, 23, etc.) when making updates to the SDK
         /// </summary>
-        public const double SDK_VERSION = 20;
+        public const int SDK_VERSION = 20;
 
         public static readonly Dictionary<string, string> MaskingMap = new Dictionary<string, string>()
         {
